try modularized adapter and train prep
